backend/requirements.txt
Flask==1.0.2
flask-restful==0.3.8
flask-cors==3.0.10
werkzeug==0.16.0
pyjwt==2.0.1
snownlp==0.12.3
jieba==0.42.1
pyyaml==5.3.1